+++ src/main/java/com/tubecentric/waldo/twitter/model/TwitterDTO.java
@@ -1,10 +1,7 @@
 package com.tubecentric.waldo.twitter.model;
 
 import lombok.Builder;
-import lombok.Singular;
 import lombok.Value;
-
-import java.util.List;
 
 @Value
 @Builder(builderClassName = "Builder")
@@ -15,6 +12,6 @@ public class TwitterDTO {
     private final String screenName;
     private final String description;
     private final String url;
-    private final Long followers;
+    private final Integer followers;
     private final TweetDTO tweet;
 }
